pyproject.toml
[build-system]
requires = ["setuptools==75.1.0"]
build-backend = "setuptools.build_meta"

[project]
name        = "backlight_control"
version     = "0.1a0"
description = "Automatic control for backlights."
requires-python = ">=3.11.0"
dependencies    = [
    "dbus-fast",
    "python-xlib",
]

[project.scripts]
backlight_control = "backlight_control.__main__:main"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle
    "F",      # Pyflakes
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "I",      # isort
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = [
    "pyotgw",
]
combine-as-imports = true
split-on-trailing-comma = false


[tool.setuptools.package-data]
"*" = ["*.xml"]